refactor!: Remove old API methods
