+++ lib/discordrb/api/client.rb
@@ -2,19 +2,19 @@
 
 require 'faraday'
 require 'faraday_middleware'
-require 'discordrb/api/client/application_command_endpoints'
-require 'discordrb/api/client/audit_log_endpoints'
-require 'discordrb/api/client/channel_endpoints'
-require 'discordrb/api/client/emoji_endpoints'
-require 'discordrb/api/client/guild_endpoints'
-require 'discordrb/api/client/guild_template_endpoints'
-require 'discordrb/api/client/interaction_endpoints'
-require 'discordrb/api/client/invite_endpoints'
-require 'discordrb/api/client/stage_instance_endpoints'
-require 'discordrb/api/client/sticker_endpoints'
-require 'discordrb/api/client/user_endpoints'
-require 'discordrb/api/client/voice_endpoints'
-require 'discordrb/api/client/webhook_endpoints'
+require 'discordrb/api/endpoints/application_command'
+require 'discordrb/api/endpoints/audit_log'
+require 'discordrb/api/endpoints/channel'
+require 'discordrb/api/endpoints/emoji'
+require 'discordrb/api/endpoints/guild'
+require 'discordrb/api/endpoints/guild_template'
+require 'discordrb/api/endpoints/interaction'
+require 'discordrb/api/endpoints/invite'
+require 'discordrb/api/endpoints/stage_instance'
+require 'discordrb/api/endpoints/sticker'
+require 'discordrb/api/endpoints/user'
+require 'discordrb/api/endpoints/voice'
+require 'discordrb/api/endpoints/webhook'
 require 'discordrb/errors'
 
 module Discordrb
