+++ lib/discordrb/api/endpoints/application_command.rb
