+++ lib/discordrb/api/endpoints/channel.rb
