+++ lib/discordrb/api/endpoints/emoji.rb
