+++ lib/discordrb/api/endpoints/guild_template.rb
