+++ lib/discordrb/api/endpoints/invite.rb
