+++ lib/discordrb/api/endpoints/stage_instance.rb
