+++ lib/discordrb/api/endpoints/sticker.rb
