+++ lib/discordrb/api/endpoints/user.rb
